Hackathons/MTC_True_Tech/iiassistant/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-leaflet==0.29.0
djangorestframework==3.14.0
ffmpeg==1.4
ffprobe==0.5
filelock==3.13.4
fsspec==2024.3.1
GDAL @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_7bpu2_kz5n/croot/gdal-split_1691162695819/work/build/swig/python
gunicorn==21.2.0
huggingface-hub==0.22.2
idna==3.6
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
mbutil==0.3.0
mkl-fft @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_af8wl3mjfu/croot/mkl_fft_1695058185839/work
mkl-random @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_0fq2s894lk/croot/mkl_random_1695059822529/work
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.3
numpy @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_dfryapvjbb/croot/numpy_and_numpy_base_1701362404676/work/dist/numpy-1.26.2-cp310-cp310-macosx_10_9_x86_64.whl#sha256=823a88c82e6cc4575319f712a9450d2dffa9fc5a077339f1d04b1df6a541b6b0
packaging==24.0
Pillow @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_18z8l26k8d/croot/pillow_1696580044342/work
psycopg2==2.9.9
PyAudio==0.2.11
pycparser==2.22
pydub==0.25.1
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
sqlparse==0.4.4
srt==3.5.3
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.9.0
urllib3==2.1.0
vosk==0.3.44
websockets==12.0
